+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/PaymentOptionsActivityTest.kt
@@ -18,6 +18,9 @@ import com.stripe.android.PaymentConfiguration
 import com.stripe.android.R
 import com.stripe.android.core.Logger
 import com.stripe.android.core.injection.DUMMY_INJECTOR_KEY
+import com.stripe.android.link.LinkPaymentLauncher
+import com.stripe.android.link.injection.FakeLinkPaymentLauncherFactory
+import com.stripe.android.link.model.AccountStatus
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodFixtures
 import com.stripe.android.paymentsheet.PaymentOptionsViewModel.TransitionTarget
@@ -41,7 +44,9 @@ import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mockito.kotlin.any
 import org.mockito.kotlin.mock
+import org.mockito.kotlin.stub
 import org.robolectric.RobolectricTestRunner
 import kotlin.test.BeforeTest
 
@@ -447,6 +452,11 @@ class PaymentOptionsActivityTest {
         val addressRepository = AddressRepository(
             ApplicationProvider.getApplicationContext<Application>().resources
         )
+        val linkPaymentLauncherFactory = FakeLinkPaymentLauncherFactory(
+            mock<LinkPaymentLauncher>().stub {
+                onBlocking { setup(any(), any()) }.thenReturn(AccountStatus.SignedOut)
+            }
+        )
         return PaymentOptionsViewModel(
             args = args,
             prefsRepositoryFactory = { FakePrefsRepository() },
@@ -463,7 +473,7 @@ class PaymentOptionsActivityTest {
                 addressRepository
             ),
             savedStateHandle = SavedStateHandle(),
-            linkPaymentLauncherFactory = mock()
+            linkPaymentLauncherFactory = linkPaymentLauncherFactory
         )
     }
 }
