+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/PaymentOptionsViewModelTestInjection.kt
@@ -1,6 +1,7 @@
 package com.stripe.android.paymentsheet
 
 import android.app.Application
+import android.content.Context
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.SavedStateHandle
 import androidx.test.core.app.ApplicationProvider
@@ -9,13 +10,16 @@ import com.stripe.android.core.injection.Injectable
 import com.stripe.android.core.injection.Injector
 import com.stripe.android.core.injection.InjectorKey
 import com.stripe.android.core.injection.WeakMapInjectorRegistry
+import com.stripe.android.link.injection.FakeLinkPaymentLauncherFactory
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.paymentsheet.analytics.EventReporter
 import com.stripe.android.paymentsheet.forms.FormViewModel
 import com.stripe.android.paymentsheet.injection.FormViewModelSubcomponent
 import com.stripe.android.paymentsheet.injection.PaymentOptionsViewModelSubcomponent
 import com.stripe.android.paymentsheet.viewmodels.BaseSheetViewModel
+import com.stripe.android.ui.core.address.AddressRepository
 import com.stripe.android.ui.core.forms.resources.LpmRepository
+import com.stripe.android.ui.core.forms.resources.StaticAddressResourceRepository
 import com.stripe.android.ui.core.forms.resources.StaticLpmResourceRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
@@ -34,6 +38,9 @@ internal open class PaymentOptionsViewModelTestInjection {
     val rule = InstantTaskExecutorRule()
 
     private val testDispatcher = UnconfinedTestDispatcher()
+    private val addressResourceRepository = StaticAddressResourceRepository(
+        AddressRepository(ApplicationProvider.getApplicationContext<Context>().resources)
+    )
 
     val eventReporter = mock<EventReporter>()
 
@@ -65,11 +72,11 @@ internal open class PaymentOptionsViewModelTestInjection {
             logger = Logger.noop(),
             injectorKey = injectorKey,
             lpmResourceRepository = StaticLpmResourceRepository(lpmRepository),
-            addressResourceRepository = mock(),
+            addressResourceRepository = addressResourceRepository,
             savedStateHandle = SavedStateHandle().apply {
                 set(BaseSheetViewModel.SAVE_RESOURCE_REPOSITORY_READY, true)
             },
-            linkPaymentLauncherFactory = mock()
+            linkPaymentLauncherFactory = FakeLinkPaymentLauncherFactory(mock())
         )
     }
 
@@ -82,7 +89,7 @@ internal open class PaymentOptionsViewModelTestInjection {
             paymentMethodCode = PaymentMethod.Type.Card.code,
             config = mock(),
             lpmResourceRepository = StaticLpmResourceRepository(lpmRepository),
-            addressResourceRepository = mock(),
+            addressResourceRepository = addressResourceRepository,
             transformSpecToElement = mock()
         )
     ) {
