+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/PaymentSheetViewModelTestInjection.kt
@@ -15,8 +15,7 @@ import com.stripe.android.core.injection.WeakMapInjectorRegistry
 import com.stripe.android.googlepaylauncher.GooglePayPaymentMethodLauncher
 import com.stripe.android.googlepaylauncher.GooglePayPaymentMethodLauncherContract
 import com.stripe.android.googlepaylauncher.injection.GooglePayPaymentMethodLauncherFactory
-import com.stripe.android.link.LinkPaymentLauncher
-import com.stripe.android.link.injection.LinkPaymentLauncherFactory
+import com.stripe.android.link.injection.FakeLinkPaymentLauncherFactory
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.StripeIntent
 import com.stripe.android.payments.paymentlauncher.StripePaymentLauncherAssistedFactory
@@ -28,7 +27,6 @@ import com.stripe.android.paymentsheet.model.StripeIntentValidator
 import com.stripe.android.paymentsheet.repositories.StripeIntentRepository
 import com.stripe.android.paymentsheet.viewmodels.BaseSheetViewModel
 import com.stripe.android.ui.core.address.AddressRepository
-import com.stripe.android.ui.core.elements.IdentifierSpec
 import com.stripe.android.ui.core.forms.resources.LpmRepository
 import com.stripe.android.ui.core.forms.resources.StaticAddressResourceRepository
 import com.stripe.android.ui.core.forms.resources.StaticLpmResourceRepository
@@ -54,7 +52,7 @@ internal open class PaymentSheetViewModelTestInjection {
     val eventReporter = mock<EventReporter>()
     private val googlePayPaymentMethodLauncherFactory = createGooglePayPaymentMethodLauncherFactory()
     private val stripePaymentLauncherAssistedFactory = mock<StripePaymentLauncherAssistedFactory>()
-    private val linkPaymentLauncherFactory = createLinkPaymentLauncherFactory()
+    private val linkPaymentLauncherFactory = FakeLinkPaymentLauncherFactory(mock())
 
     private lateinit var injector: Injector
 
@@ -75,19 +73,6 @@ internal open class PaymentSheetViewModelTestInjection {
                 val googlePayPaymentMethodLauncher = mock<GooglePayPaymentMethodLauncher>()
                 readyCallback.onReady(true)
                 return googlePayPaymentMethodLauncher
-            }
-        }
-
-    private fun createLinkPaymentLauncherFactory() =
-        object : LinkPaymentLauncherFactory {
-            override fun create(
-                merchantName: String,
-                customerEmail: String?,
-                customerPhone: String?,
-                customerName: String?,
-                shippingValues: Map<IdentifierSpec, String?>?
-            ): LinkPaymentLauncher {
-                return mock()
             }
         }
